Organizing structure of header
Modularizando iconos, quitando info redundante, folder header unicamente para testeo
