+++ client/src/app/header/page.tsx
@@ -1,0 +1,5 @@
+import Header from '@/components/common/Header';
+
+export default function HeaderTest() {
+  return <Header />;
+}
