+++ client/src/assets/Icons/EditIcon.tsx
@@ -1,0 +1,21 @@
+interface Props {
+  className: string;
+}
+
+export default function EditIcon({ className }: Props) {
+  return (
+    <svg
+      className={className}
+      viewBox='0 0 24 24'
+      xmlns='http://www.w3.org/2000/svg'
+      fill='none'
+      stroke='currentColor'
+      strokeWidth='2'
+      strokeLinecap='round'
+      strokeLinejoin='round'
+    >
+      {' '}
+      <path d='M17 3a2.828 2.828 0 1 1 4 4L7.5 20.5 2 22l1.5-5.5L17 3z' />
+    </svg>
+  );
+}
