+++ client/src/assets/Icons/IconLogoFinder.tsx
@@ -1,0 +1,25 @@
+interface Props {
+  className: string;
+}
+
+export default function IconLogoFinder({ className }: Props) {
+  return (
+    <svg
+      className={className}
+      viewBox='0 0 24 24'
+      strokeWidth='2'
+      stroke='currentColor'
+      fill='none'
+      strokeLinecap='round'
+      strokeLinejoin='round'
+    >
+      {' '}
+      <path stroke='none' d='M0 0h24v24H0z' />{' '}
+      <path d='M3 19a9 9 0 0 1 9 0a9 9 0 0 1 9 0' />{' '}
+      <path d='M3 6a9 9 0 0 1 9 0a9 9 0 0 1 9 0' />{' '}
+      <line x1='3' y1='6' x2='3' y2='19' />{' '}
+      <line x1='12' y1='6' x2='12' y2='19' />{' '}
+      <line x1='21' y1='6' x2='21' y2='19' />
+    </svg>
+  );
+}
