+++ client/src/assets/Icons/SearchIcon.tsx
@@ -1,0 +1,21 @@
+interface Props {
+  className: string;
+}
+
+export default function SearchIcon({ className }: Props) {
+  return (
+    <svg
+      className={className}
+      viewBox='0 0 24 24'
+      fill='none'
+      stroke='currentColor'
+      strokeWidth='2'
+      strokeLinecap='round'
+      strokeLinejoin='round'
+    >
+      {' '}
+      <circle cx='11' cy='11' r='8' />{' '}
+      <line x1='21' y1='21' x2='16.65' y2='16.65' />
+    </svg>
+  );
+}
